Refactored tests based on test reason
