+++ tests/AoPeas.Tests/Fixtures/TestService.cs
@@ -12,7 +12,8 @@ public interface ITestService
 
 public class TestService : ITestService
 {
-    [EnableProxyLogging]
+    [NoAdvice]
+    [AddTen]
     public int GetIncrement(int a) => a + 1;
     [NotImplementedPointcut]
     public int GetSum(int a, int b) => a + b;
@@ -28,6 +29,13 @@ public class NoAdviceAttribute : PointcutAttribute { }
 public class NoAdvice : IAdvice<NoAdviceAttribute>
 {
     public object? Apply(MethodInvocationDetails invocationDetails) => invocationDetails.Next();
+}
+
+public class AddTenAttribute : PointcutAttribute { }
+
+public class AddTen : IAdvice<AddTenAttribute>
+{
+    public object? Apply(MethodInvocationDetails invocationDetails) => (invocationDetails.Next() as int?) + 10;
 }
 
 public class EnableProxyLoggingAttribute : PointcutAttribute { }
